Grid user-interactivity
The user can now click on the node (tile) in the grid.

If the node is not blocked (white) and it is clicked it then becomes blocked and turns red.

If the node is blocked (so red) and the user clicks it, it now becomes unblocked and turns back to white