Settings update
The new functionality is more dynamic and easily handled. Prior to this commit, the user would have to set the amount of rows and columns the grid would contain. They would also also have to set the size of each square and modify multiple values from different classes.

The user sets the Settings width, height and node size, this now means the grid is drawn in relation to the width and height and node size. set by the user, for example.

If width equals 200, and the node size is 20 then the amount of rows is equal to 200/20 = 10. With this update we can also set how large or small we want each tile to be in the grid.

We had to add a value of 1 to the screen width and height due to how the graphics are drawn.