Selection method added
The method for selecting an Agent from the population has been added. Currently it just creates an array "matingPool", pushes agents x number of times based on fitness * 10, picks a random agent from the mating pool and sets this to be the child. We return the child agent and set the current agents array at index i to the new child.